+++ webapp/src/pages/dashboard/offer/obiz/[id]/index.tsx
@@ -62,6 +62,10 @@ export default function OfferObizPage({ offer_id }: OfferObizPageProps) {
     return getItemsConditionBlocks(offer.kind) as StackItem[];
   }, [offer]);
 
+  const onRedirectPayment = () => {
+    onCloseOrderProcessModal();
+  };
+
   useEffect(() => {
     const mutateData = async () => {
       await increaseNbSeen({ offer_id: parseInt(offer_id) });
@@ -299,6 +303,7 @@ export default function OfferObizPage({ offer_id }: OfferObizPageProps) {
         <ObizOrderProcessModal
           isOpen={isOpenOrderProcessModal}
           onClose={onCloseOrderProcessModal}
+          onRedirectPayment={onRedirectPayment}
           offerId={parseInt(offer_id)}
         />
       </Flex>